+++ src/index.ts
@@ -1,0 +1,3 @@
+export { default as StarRatingInput, Props as StarRatingInputProps } from './StarRatingInput'
+export { default as StarRating, Props as StarRatingProps } from './StarRating'
+export { default as css } from './css'