+++ src/insert-css.d.ts
@@ -1,0 +1,4 @@
+declare module 'insert-css' {
+    const insertCss: (css: string) => void
+    export default insertCss
+}